refactor: station 관련 dto 패키징 변경
