+++ src/main/java/wooteco/subway/station/controller/StationController.java
@@ -13,8 +13,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import wooteco.subway.station.controller.dto.StationCreateDto;
-import wooteco.subway.station.controller.dto.StationDto;
+import wooteco.subway.station.service.dto.StationCreateDto;
+import wooteco.subway.station.service.dto.StationDto;
 import wooteco.subway.station.controller.dto.StationRequest;
 import wooteco.subway.station.controller.dto.StationResponse;
 import wooteco.subway.station.service.StationService;
