+++ src/main/java/wooteco/subway/station/controller/dto/StationRequest.java
@@ -2,6 +2,7 @@ package wooteco.subway.station.controller.dto;
 
 import javax.validation.constraints.Pattern;
 import javax.validation.constraints.Size;
+import wooteco.subway.station.service.dto.StationCreateDto;
 
 public class StationRequest {
